Generalized platform cards (WIP)
