+++ src/components/[guild]/RolePlatforms/components/PlatformCard/components/TelegramCard.tsx
@@ -1,8 +1,13 @@
 import useIsTGBotIn from "components/create-guild/TelegramGroup/hooks/useIsTGBotIn"
+import { PropsWithChildren } from "react"
 import { useRolePlatform } from "../../RolePlatformProvider"
 import PlatformCard from "../PlatformCard"
 
-const TelegramCard = ({ onRemove }) => {
+type Props = {
+  onRemove?: () => void
+}
+
+const TelegramCard = ({ onRemove, children }: PropsWithChildren<Props>) => {
   const { guildPlatform } = useRolePlatform()
   const {
     data: { groupIcon, groupName },
@@ -14,7 +19,9 @@ const TelegramCard = ({ onRemove }) => {
       onRemove={onRemove}
       imageUrl={groupIcon || "/default_telegram_icon.png"}
       name={groupName || ""}
-    />
+    >
+      {children}
+    </PlatformCard>
   )
 }
